First steps for refactoring of layer configuration class heirarchy
