+++ deeplearning4j-nn/src/main/java/org/deeplearning4j/nn/layers/BaseOutputLayer.java
@@ -337,7 +337,10 @@ public abstract class BaseOutputLayer<LayerConfT extends org.deeplearning4j.nn.c
             solver = new Solver.Builder().configure(conf()).listeners(getListeners()).model(this).build();
             //Set the updater state view array. For MLN and CG, this is done by MultiLayerUpdater and ComputationGraphUpdater respectively
             Updater updater = solver.getOptimizer().getUpdater();
-            int updaterStateSize = (int) conf().getLayer().getIUpdater().stateSize(numParams());
+            int updaterStateSize = 0;
+            for(String s : paramTable().keySet()){
+                updaterStateSize += (int) conf().getLayer().getIUpdaterByParam(s).stateSize(numParams());
+            }
             if (updaterStateSize > 0)
                 updater.setStateViewArray(this, Nd4j.createUninitialized(new int[] {1, updaterStateSize}, Nd4j.order()),
                                 true);
